feat: exercise 4.21 - restrict blog deletion to the creator user

+++ controllers/blogs.js
@@ -52,8 +52,24 @@ blogsRouter.post("/", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const blog = await Blog.findById(request.params.id);
 
+  if (!blog) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({
+      error: "only the creator can delete this blog",
+    });
+  }
+
+  await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
